Start implementing the new batcher for GL2

+++ kglt/material.h
@@ -112,7 +112,7 @@ public:
     Colour specular() const { return specular_; }
     float shininess() const { return shininess_; }
 
-    int32_t texture_unit_count() const { return (int) texture_units_.size(); }
+    uint32_t texture_unit_count() const { return texture_units_.size(); }
     TextureUnit& texture_unit(uint32_t index) { return texture_units_.at(index); }
 
     void update(double dt) {
